% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apodos.R
\docType{data}
\name{apodos}
\alias{apodos}
\title{apodos}
\format{A data frame with 5362 rows and 4 variables:
\describe{
  \item{\code{id_unico_ruvte}}{identificador RUVTE}
  \item{\code{nombre}}{Nombre y Apellido oficial}
  \item{\code{apellido_materno}}{Apellido Materno}
  \item{\code{apodo}}{Apodo}
}}
\usage{
apodos
}
\description{
se detalla el listado de apodos.
Recurso original: http://datos.jus.gob.ar/dataset/registro-unificado-de-victimas-del-terrorismo-de-estado-ruvte
}
\keyword{datasets}
