% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parque_de_la_memoria.R
\docType{data}
\name{parque_de_la_memoria}
\alias{parque_de_la_memoria}
\title{parque_de_la_memoria}
\format{Data frame con 8751 filas y 30 variables:
\describe{
  \item{\code{id}}{character id del caso. Construido a partir del id utilizado en la página}
  \item{\code{link}}{character url al registro de parquedelamemoria.org}
  \item{\code{nombre}}{character }
  \item{\code{edad}}{character }
  \item{\code{embarazada}}{character Indica si la víctima se encontraba embarazada al momento del hecho}
  \item{\code{estado}}{character Se distinguen dos tipos de víctimas del accionar represivo del Estado: por un lado, detenidos-desaparecidos (aquellas personas cuyos restos no han sido hallados aún) y, por el otro, asesinados (los casos donde sí se ha identificado sus restos y el Registro Civil ha expedido un certificado de defunción).}
  \item{\code{ano_en_monumento}}{character Año en que se incorpora al monumento del parque}
  \item{\code{domicilios}}{character }
  \item{\code{estado_civil}}{character }
  \item{\code{estudios}}{character }
  \item{\code{fecha_de_nacimiento}}{character }
  \item{\code{fecha_de_secuestro}}{character }
  \item{\code{lugar_de_nacimiento}}{character }
  \item{\code{lugar_de_secuestro}}{character }
  \item{\code{nacionalidad}}{character }
  \item{\code{ocupaciones}}{character }
  \item{\code{sexo}}{character }
  \item{\code{victimas_simultaneas}}{character Personas detenidas y/o asesinadas en el mismo operativo represivo que la víctima que se consigna en el Registro.}
  \item{\code{foto}}{character }
  \item{\code{apodos}}{character }
  \item{\code{cantidad_de_hijos}}{character }
  \item{\code{victimas_relacionadas}}{character }
  \item{\code{casos_relacionados}}{list }
  \item{\code{militancia}}{character }
  \item{\code{articulos_periodisticos}}{character }
  \item{\code{datos_de_la_identificacion}}{character }
  \item{\code{fecha_de_asesinado}}{character }
  \item{\code{lugar_de_asesinato}}{character }
  \item{\code{observaciones}}{character }
  \item{\code{datos_familiares}}{character }
}}
\usage{
parque_de_la_memoria
}
\description{
Dataset elaborado a partir de los registros de
 http://basededatos.parquedelamemoria.org.ar/registros/
}
\keyword{datasets}
