% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_accionar_represivo_ilegal.R
\docType{data}
\name{victimas_accionar_represivo_ilegal}
\alias{victimas_accionar_represivo_ilegal}
\title{victimas_accionar_represivo_ilegal}
\format{A data frame with 8753 rows and 20 variables:
\describe{
  \item{\code{id_unico_ruvte}}{character número de identificación del caso individual}
  \item{\code{anio_denuncia}}{double año de alta original de la denuncia en los registros que hoy constituyen el Ruvte}
  \item{\code{tipificacion_ruvte}}{character tipificación principal del caso según el Ruvte (víctima de desaparición forzada
   o víctima de asesinato), a la que se agregan observaciones que en algunos casos especifican la fuente, y en otros,
    actualizaciones de estado (por ejemplo, exhumados e identificados sus restos, entre otras)}
  \item{\code{apellido_paterno_nombres}}{character apellido paterno y nombres de pila de la víctima}
  \item{\code{apellido_materno}}{character apellido materno}
  \item{\code{apellido_casada}}{character apellido de casada}
  \item{\code{edad_al_momento_del_hecho}}{character edad al momento del hecho (al momento de su deceso, o, en su defecto,
   de su “desaparición”)}
  \item{\code{documentos}}{character tipo y número de documento}
  \item{\code{anio_nacimiento}}{character año de nacimiento}
  \item{\code{provincia_nacimiento}}{character provincia de nacimiento (solo para el caso de ciudadanos nacidos en Argentina)}
  \item{\code{pais_nacimiento}}{character país de nacimiento}
  \item{\code{nacionalidad}}{character nacionalidad de la víctima}
  \item{\code{embarazo}}{character datos sobre el embarazo de la víctima y sobre el destino del niño nacido en cautiverio}
  \item{\code{fecha_detencion_secuestro}}{character fecha de detención-secuestro de la víctima}
  \item{\code{lugar_detencion_secuestro}}{character lugar de detención-secuestro de la víctima}
  \item{\code{fecha_asesinato_o_hallazgo_de_restos}}{character fecha de asesinato, fallecimiento o hallazgo de los restos de la víctima}
  \item{\code{lugar_asesinato_o_hallazgo_de_restos}}{character lugar de asesinato, fallecimiento o hallazgo de los restos de la víctima}
  \item{\code{fotografia}}{character información sobre si obra fotografía de la víctima en base de datos}
  \item{\code{provincia_nacimiento_indec_id}}{character provincia de nacimiento (solo para el caso de ciudadanos nacidos en Argentina), según la codificación de provincia implementada por INDEC}
  \item{\code{pais_nacimiento_indec_id}}{character país de nacimiento, según la codificación de país implementada por INDEC.}
}}
\usage{
victimas_accionar_represivo_ilegal
}
\description{
Listado de víctimas del accionar represivo ilegal.
Recurso original: http://datos.jus.gob.ar/dataset/registro-unificado-de-victimas-del-terrorismo-de-estado-ruvte
}
\keyword{datasets}
